+++ backend/src/uploads/index.js
@@ -1,0 +1,13 @@
+import axios from 'axios';
+import fs from 'fs';
+import FormData from 'form-data';
+
+import { getFile, uploadFile } from './endpoints.js';
+
+export default function useUploads(app) {
+  app.post('/upload', uploadFile);
+
+  app.get('/upload', getFile);
+
+  return app;
+}
